refactor: narrow workflow virtualization scope
